movies_backend/requirements.txt
django-admin
psycopg2
environ
djangorestframework
requests
bs4
re